Conflicts:
	datadriven/SConscript

+++ pde/src/sgpp/pde/application/PoissonEquationSolver.cpp
@@ -32,9 +32,9 @@ PoissonEquationSolver::~PoissonEquationSolver() {
   }
 }
 
-void PoissonEquationSolver::constructGrid(base::BoundingBox& BoundingBox, int level) {
+void PoissonEquationSolver::constructGrid(base::BoundingBox& BoundingBox, size_t level) {
   this->dim = BoundingBox.getDimensions();
-  this->levels = level;
+  this->levels = static_cast<int>(level);
 
   this->myGrid = new base::LinearBoundaryGrid(BoundingBox);
